+++ src/routing/definition/create-route-definition.ts
@@ -1,8 +1,8 @@
 import type { HttpMiddleware, HttpRequest } from '@/Http';
 import { resolveRouteOptions } from '@/routing/definition/resolve-route-options';
-import type { HttpMethod } from '@/routing/http-method.type';
 import type { RouteOptions } from '@/routing/route-options';
 import type { RouterMethod } from '@/routing/router-method';
+import type { HttpMethod } from '@/routing/verb/http-method.type';
 import type { Request } from 'koa';
 import type { RouteDefinition } from './route-definition';
