+++ src/routing/deprecated-decorator/route.ts
@@ -1,6 +1,6 @@
 import type { HttpMiddleware } from '@/Http';
-import type { HttpMethod } from '@/routing/http-method.type';
 import type { RouteOptions } from '@/routing/route-options';
+import type { HttpMethod } from '@/routing/verb/http-method.type';
 import { createLegacyRouteDecorator } from './legacy-router';
 
 /**
